+++ evennia/server/inputfuncs.py
@@ -156,6 +156,7 @@ _CLIENT_OPTIONS = (
     "RAW",
     "NOCOLOR",
     "NOGOAHEAD",
+    "LOCALECHO",
 )
 
 
@@ -180,6 +181,7 @@ def client_options(session, *args, **kwargs):
         inputdebug (bool): Debug input functions
         nocolor (bool): Strip color
         raw (bool): Turn off parsing
+        localecho (bool): Turn on server-side echo (for clients not supporting it)
 
     """
     old_flags = session.protocol_flags
@@ -239,6 +241,8 @@ def client_options(session, *args, **kwargs):
             flags["RAW"] = validate_bool(value)
         elif key == "nogoahead":
             flags["NOGOAHEAD"] = validate_bool(value)
+        elif key == "localecho":
+            flags["LOCALECHO"] = validate_bool(value)
         elif key in (
             "Char 1",
             "Char.Skills 1",